Alternative approach for ui:options:label = false
